adjusted seed data, updated resolver with cats data

+++ server/schemas/typeDefs.js
@@ -11,7 +11,7 @@ const typeDefs = gql`
   type Cat {
     _id: ID
     catName: String
-    catAge: Int
+    catAge: Float
     catSpecies: String
     catBio: String
     userId: ID
